DEV: Extract upload downsizing logic
